Fix JPMML compile problem in 1.2.13

+++ app/oryx-app-mllib/src/test/java/com/cloudera/oryx/app/batch/mllib/kmeans/KMeansHyperParamTuningIT.java
@@ -80,7 +80,7 @@ public final class KMeansHyperParamTuningIT extends AbstractKMeansIT {
     ClusteringModel clusteringModel = (ClusteringModel) rootModel;
 
     // Should have picked highest k
-    assertEquals(100, clusteringModel.getNumberOfClusters().intValue());
+    assertEquals(100, clusteringModel.getNumberOfClusters());
   }
 
 }
